put all genre and music_albums in separate mdoules
